fixed colors.c and controllers

+++ main/model/color.h
@@ -24,7 +24,7 @@
      * If number is greater than range2, then the lightest shade of color
      * is chosen.
      */
-    void scale_color(enum Color colorVal, int value, int maxVal, int minVal, int x, int y);
+    void scale_color(enum Color colorVal, double value, int maxVal, int minVal, int x, int y);
 
     void update_color();
